scripts/runAcceptanceTests.sh: fix Python syntax error
Python 3 needs the parens to call print, and recent Python 2s don't barf
on it.

Another option might be using `date` instead of Python, but turns out
subsecond precision for `date` is a GNU extension, so not portable to
macOS. Padding the timestamp with zeros (like `$(date +%s)000`) won't
work because we actually _need_ the correct subsecond resolution
timestamp.