scripts/runAcceptanceTests.sh: fix curl to Zipkin-in-Docker
On CircleCI 2.0 remote docker containers' ports are not exposed to the
test container (since they're not running on the same Docker network).
According to
https://circleci.com/docs/2.0/building-docker-images/#accessing-services
one way is to run _another_ container on she same network, but that
would require knowing the name of the network (which is defined by
`docker-compose` implicitly), which is more than what we want to know
here.

Luckily we have `curl` inside the Zipkin image, so we might as well use
it and execute `curl` in there.